Server implementation for importing

+++ client/src/main/scala/com/example/rest2/AjaxCall.scala
@@ -124,7 +124,7 @@ object AjaxCall extends IAjaxCall {
     }
     req.open(method, url, true)
     req.responseType = "text" // responseType
-    req.timeout = timeout.toMillis
+    req.timeout = timeout.toMillis.toDouble
     req.withCredentials = withCredentials
     headers.foreach(x => req.setRequestHeader(x._1, x._2))
     Init.clientid.map( x => req.setRequestHeader("x-clientid", x) )
